fix: Address code review feedback on syntax highlighting PR
- Remove dead _base_style parameter from highlight_code and fix doc comment
- Move messages_mermaid inline tests to separate messages_mermaid_tests.rs file
- Remove duplicate test_parse_code_fence_produces_code_block in mermaid.rs
- Add test for CodeBlock-first message timestamp gutter (issue 1 coverage)
- Add test for code line truncation at content_width (issue 2 coverage)
- Add highlight module to mod.rs doc comment and change from pub to private
